@@ -1,10 +1,10 @@
 name: users_client
-description: Orion Users Client
+description: Orion Users Client. This a dart abstraction to Orion Users service
 version: 0.1.0
 homepage: https://github.com/orion-services/users_client
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: ^2.17.5
 
 dependencies:
   prompts: ^2.0.0
